+++ components/PolarAreaChart.js
@@ -1,5 +1,5 @@
 import ColorItem from './ColorItem';
-import { prettyNumber } from '../util';
+import { prettyNumber, clamp } from '../util';
 import { COLORS } from '../util/constants';
 
 function getCoordinatesForPercent(percent, size, radii) {
@@ -9,15 +9,17 @@ function getCoordinatesForPercent(percent, size, radii) {
   };
 }
 
-export default function PieChart({
+export default function PolarAreaChart({
   data,
   colors,
-  cutout = 50,
+  size = 100,
+  radii = size / 2,
+  width = '100%'
 }) {
-  const size = 100;
-  const radii = 50;
   const values = Object.values(data);
   const total = values.reduce((acc, current) => acc + current, 0);
+  const percentages = values.map(p => p * 1 / total);
+  const maxPercentage = Math.max(...percentages);
 
   if (total === 0) {
     return (
@@ -29,11 +31,12 @@ export default function PieChart({
 
   let acc = 0;
 
-  const items = Object.entries(data).map(([key, value], index) => {
-    const percent = value * 1 / total;
-    const start = getCoordinatesForPercent(acc, size, radii);
-    const end = getCoordinatesForPercent(acc += percent, size, radii);
-    const f = percent > .5 ? 1 : 0;
+  const items = Object.entries(data).map(([key, value], i) => {
+    const percent = clamp(.1, value * 1 / total, 1);
+    const radius = percent * radii / maxPercentage;
+    const start = getCoordinatesForPercent(acc, size, radius);
+    const end = getCoordinatesForPercent(acc += 1 / values.length, size, radius);
+    const f = 1 / values.length > .5 ? 1 : 0;
 
     return {
       key,
@@ -41,41 +44,32 @@ export default function PieChart({
       d: `
         M ${size/2} ${size/2}
         L ${start.x} ${start.y}
-        A ${radii} ${radii} 0 ${f} 1 ${end.x} ${end.y}
+        A ${radius} ${radius} 0 ${f} 1 ${end.x} ${end.y}
+        Z
       `,
-      color: colors?.[key] ?? COLORS[index]
-    };
+      color: colors?.[key] ?? COLORS[i]
+    }
   });
 
   return (
     <div className='root'>
       <div className="chartWrapper">
         <svg viewBox={`0 0 ${size} ${size}`}>
-          {items.length === 1 ? (
+          {new Array(5).fill(0).map((_, i, arr) => (
             <circle
-              fill={items[0].color}
-              r={size/2}
+              key={i}
+              r={radii / arr.length * (i+1)}
               cx={size/2}
               cy={size/2}
             />
-          ) : (
-            items.map((item, i) => (
-              <path
-                key={i}
-                d={item.d}
-                fill={item.color}
-              />
-            ))
-          )}
-
-          {cutout && (
-            <circle
-              fill='var(--color-background)'
-              r={(cutout/2)}
-              cx={size/2}
-              cy={size/2}
+          ))}
+          {items.map((item, i) => (
+            <path
+              key={item.key}
+              d={item.d}
+              fill={item.color}
             />
-          )}
+          ))}
         </svg>
       </div>
       <div className='info'>
@@ -98,6 +92,15 @@ export default function PieChart({
         }
         svg {
           transform: rotate(90deg);
+        }
+        circle {
+          fill: none;
+          stroke: var(--gps-border-color);
+        }
+        path {
+          stroke: var(--color-background);
+          stroke-width: 1.5;
+          stroke-linejoin: round;
         }
         .info {
           font-size: 12px;