+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.6.0] - 2021-04-15
+#### Added
+- components: polar area chart
+- display version number on footer
+#### Changed
+- user: use PolarAreaChart instead of RepoCard
+- components: pie chart: use path instead of stroke-dasharray
+- constants: update chart colors
+
 ## [0.5.0] - 2021-04-13
 #### Added
 - user: calendar line chart
@@ -58,6 +67,7 @@ Initial release
   - languages per repo
   - commits per repo (top 10)
 
+[0.6.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/noeldelgado/gh-profile-stats/compare/v0.2.1...v0.3.0